refactor: service command dto 패키징
