+++ src/main/java/project/volunteer/domain/sehedule/api/dto/request/ScheduleUpsertRequest.java
@@ -10,7 +10,7 @@ import org.hibernate.validator.constraints.Range;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
-import project.volunteer.domain.sehedule.application.dto.ScheduleUpsertCommand;
+import project.volunteer.domain.sehedule.application.dto.command.ScheduleUpsertCommand;
 
 @Getter
 @Setter
