+++ src/main/java/project/volunteer/domain/sehedule/application/ScheduleCommandService.java
@@ -5,13 +5,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.volunteer.domain.recruitment.domain.Recruitment;
-import project.volunteer.domain.sehedule.application.dto.RegularScheduleCreateCommand;
+import project.volunteer.domain.sehedule.application.dto.command.RegularScheduleCreateCommand;
 import project.volunteer.domain.sehedule.application.timetableCreator.RepeatTimetableCreateProvider;
 import project.volunteer.domain.sehedule.application.timetableCreator.RepeatTimetableCreator;
 import project.volunteer.domain.sehedule.repository.ScheduleJdbcRepository;
 import project.volunteer.domain.sehedule.repository.ScheduleRepository;
 import project.volunteer.domain.sehedule.domain.Schedule;
-import project.volunteer.domain.sehedule.application.dto.ScheduleUpsertCommand;
+import project.volunteer.domain.sehedule.application.dto.command.ScheduleUpsertCommand;
 import project.volunteer.global.common.component.IsDeleted;
 import project.volunteer.global.common.component.Timetable;
 import project.volunteer.global.error.exception.BusinessException;
