+++ src/test/java/project/volunteer/domain/sehedule/application/ScheduleCommandUseCaseTest.java
@@ -18,8 +18,8 @@ import project.volunteer.domain.recruitment.domain.VolunteerType;
 import project.volunteer.domain.recruitment.domain.VolunteeringCategory;
 import project.volunteer.domain.recruitment.domain.VolunteeringType;
 import project.volunteer.domain.recruitment.domain.Week;
-import project.volunteer.domain.sehedule.application.dto.RegularScheduleCreateCommand;
-import project.volunteer.domain.sehedule.application.dto.ScheduleUpsertCommand;
+import project.volunteer.domain.sehedule.application.dto.command.RegularScheduleCreateCommand;
+import project.volunteer.domain.sehedule.application.dto.command.ScheduleUpsertCommand;
 import project.volunteer.domain.sehedule.domain.Schedule;
 import project.volunteer.global.common.component.Address;
 import project.volunteer.global.common.component.Coordinate;
